Refactor: change test to new private function, remove ghost cost and update comments
