+++ .github/workflows/test.yml
@@ -18,7 +18,7 @@ jobs:
     steps:
     - uses: actions/checkout@v4
     - name: Run clippy
-      run: cargo clippy --all-targets -- -D clippy::all -D clippy::nursery
+      run: cargo clippy --all-targets --all-features -- -D warnings -D clippy::all -D clippy::nursery
 
   cargo-fmt:
     runs-on: ubuntu-20.04
