+++ Cargo.toml
@@ -17,8 +17,6 @@ borsh = "0.10.1"
 hex = "0.4"
 serde = "1.0"
 
-
-
 [dev-dependencies]
 near-sdk = { version = "5.1.0", features = ["unit-testing"] }
 near-workspaces = { version = "0.10.0", features = ["unstable"] }
