+++ justfile
@@ -8,5 +8,4 @@ build:
 
 clippy:
     echo "Running clippy"
-    cargo clippy --all-targets --all-features -- -D warnings
-    cargo clippy --all-targets -- -D clippy::all -D clippy::nursery
+    cargo clippy --all-targets --all-features -- -D warnings -D clippy::all -D clippy::nursery
